requirements.txt
asgiref==3.2.3
certifi==2018.8.24
chardet==3.0.4
Deprecated==1.2.7
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
flatten-dict==0.2.0
idna==2.6
pathlib2==2.3.5
pbr==5.4.4
pprint==0.1
PyGithub==1.45
PyJWT==1.7.1
pytz==2019.3
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
stevedore==1.31.0
urllib3==1.24.1
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
